LayoutTests/editing/spelling/grammar-arrow-upstream-into-multiword-phrase-expected.txt

This test verifies that the grammar correction UI does not appear when arrowing into a multi-word grammar marker from upstream (left). This test requires WebKitTestRunner to run.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS internals.hasGrammarMarker(9, 7) became true
PASS internals.isAlternativeTextUIActive() is false
PASS successfullyParsed is true

TEST COMPLETE
